+++ lib/src/common/extensions/context.dart
@@ -6,6 +6,8 @@ extension ISpectContextExtension on BuildContext {
   /// `theme` returns the current `ThemeData` of the `BuildContext`.
   ThemeData get ispectTheme => Theme.of(this);
 
+  bool get isDarkMode => ispectTheme.brightness == Brightness.dark;
+
   /// Returns the current `ISpectAppLocalizations` of the `BuildContext`.
   ISpectGeneratedLocalization get ispectL10n => ISpectLocalization.of(this);
 }