+++ lib/src/features/view/talker_view.dart
@@ -64,8 +64,7 @@ class _TalkerViewState extends State<TalkerView> {
           builder: (context, child) => TalkerBuilder(
             talker: widget.talker,
             builder: (context, data) {
-              final filtredElements =
-                  data.where((e) => _controller.filter.filter(e)).toList();
+              final filtredElements = data.where((e) => _controller.filter.filter(e)).toList();
               final titles = data.map((e) => e.title).toList();
               final uniqTitles = titles.toSet().toList();
 
@@ -89,7 +88,7 @@ class _TalkerViewState extends State<TalkerView> {
                       // ISpectToaster.showInfoToast(context, title: context.ispectL10n.app_data);
                     },
                     onToggleTitle: _onToggleTitle,
-                    isDark: widget.options.themeMode == ThemeMode.dark,
+                    isDark: context.isDarkMode,
                   ),
                   const SliverToBoxAdapter(child: SizedBox(height: 8)),
                   SliverList(
@@ -105,7 +104,7 @@ class _TalkerViewState extends State<TalkerView> {
                           onCopyTap: () => _copyTalkerDataItemText(data),
                           expanded: _controller.expandedLogs,
                           color: getTypeColor(
-                            isDark: widget.options.themeMode == ThemeMode.dark,
+                            isDark: context.isDarkMode,
                             key: data.title,
                           ),
                         );
@@ -133,8 +132,7 @@ class _TalkerViewState extends State<TalkerView> {
     List<TalkerData> filtredElements,
     int i,
   ) {
-    final data = filtredElements[
-        _controller.isLogOrderReversed ? filtredElements.length - 1 - i : i];
+    final data = filtredElements[_controller.isLogOrderReversed ? filtredElements.length - 1 - i : i];
     return data;
   }
 
@@ -192,12 +190,8 @@ class _TalkerViewState extends State<TalkerView> {
           ),
           TalkerActionItem(
             onTap: _toggleLogsExpanded,
-            title: _controller.expandedLogs
-                ? context.ispectL10n.collapseLogs
-                : context.ispectL10n.expandLogs,
-            icon: _controller.expandedLogs
-                ? Icons.visibility_outlined
-                : Icons.visibility_off_outlined,
+            title: _controller.expandedLogs ? context.ispectL10n.collapseLogs : context.ispectL10n.expandLogs,
+            icon: _controller.expandedLogs ? Icons.visibility_outlined : Icons.visibility_off_outlined,
           ),
           TalkerActionItem(
             onTap: _cleanHistory,